Adds getComponentClassName Logic
Updates getClassName method of Mixin to call a getComponentClassName
method in the main component if one exists. The return value from that
method is then used as the base and the argument of getClassName and
the className prop are merged into this return value.

Moves logic of convertStringToClassSetObject to a utility function.
Creates a mergeClassSetObject utility function.

Adds new tests.

Updates README.md and example.